Style changes + TokenAmount refactor
