+++ frontend/src/components/AddLiquidity.tsx
@@ -5,7 +5,7 @@ import {selectUserAccountAddress} from '../redux/reducers/user';
 import {selectTokenList} from '../redux/reducers/tokens';
 import {setAccountAddress, setTokenList, setFirstToken, setSecondToken} from '../redux/actions';
 
-import TokenAmount from './TokenAmount';
+import TokenAmount from './TokenAmount/TokenAmount';
 
 /* eslint-dsiable */
 // @ts-ignore
