+++ frontend/src/components/CreatePair.scss
@@ -34,20 +34,20 @@
     margin-top: 25px;
 
     .CreatePair-button {
-      background-color: #00c3b4;
+      background-color: #a30786;
       color: #ffffff;
       padding: 10px 15px;
       width: 100%;
       height: 100%;
       font-size: 18px;
-      border: 1px solid #00c3b4;
+      border: 1px solid #a30786;
       border-radius: 12px;
       text-transform: uppercase;
       font-weight: 600;
       cursor: pointer;
 
       &:hover {
-        background-color: #00ab9e;
+        background-color: #940678;
       }
     }
   }
@@ -87,7 +87,7 @@
       }
 
       &:hover {
-        border: 2px solid #a30786;
+        border: 2px solid #940678;
       }
     }
   }
