+++ frontend/src/components/NavigationBar.scss
@@ -20,22 +20,6 @@
       border-bottom: 2px solid #a30786;
     }
 
-    .Navbar-connect-wallet {
-      background-color: #00c3b4;
-      color: #ffffff;
-      padding: 10px 15px;
-      height: 100%;
-      font-size: 18px;
-      border: 1px solid #00c3b4;
-      text-transform: uppercase;
-      font-weight: 600;
-      cursor: pointer;
-
-      &:hover {
-        background-color: #00ab9e;
-      }
-    }
-
     a {
       margin: 0 8px;
       color: black;
