+++ frontend/src/components/SwapComponent/SwapModal.scss
@@ -50,7 +50,8 @@
       font-size: 20px !important;
       padding: 20px;
       margin-top: 10px;
-      border: 2px solid lightgray;
+      border-radius: 12px;
+      border: 2px solid #a30786;
     }
   }
 
@@ -58,20 +59,20 @@
     margin-top: 25px;
 
     .SwapComponent-swap-modal-button {
-      background-color: #00c3b4;
+      background-color: #a30786;
       color: #ffffff;
       padding: 10px 15px;
       width: 100%;
       height: 100%;
       font-size: 18px;
-      border: 1px solid #00c3b4;
+      border: 1px solid #a30786;
       border-radius: 12px;
       text-transform: uppercase;
       font-weight: 600;
       cursor: pointer;
 
       &:hover {
-        background-color: #00ab9e;
+        background-color: #940678;
       }
     }
   }
