--- frontend/src/components/TokenAmount.scss
@@ -1,117 +1,0 @@
-.TokenAmount {
-  display: flex;
-  flex-direction: column;
-  border: 2px solid lightgray;
-  border-radius: 12px;
-  padding: 16px;
-
-  .TokenAmount-title {
-    margin: 0 0 12px 0;
-    font-size: 16px;
-    text-align: left;
-  }
-
-  .TokenAmount-content {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
-
-  .TokenAmount-amount {
-    background-color: transparent;
-    font-size: 24px;
-    border: none;
-    outline: none;
-  }
-
-  .TokenAmount-token-select {
-    background-color: transparent;
-    background-repeat: no-repeat;
-    border: none;
-    border-radius: 12px;
-    outline: none;
-    cursor: pointer;
-    font-size: 16px;
-    font-weight: 600;
-    padding: 0 1rem;
-    height: 2.2rem;
-    user-select: none;
-
-    &:hover {
-      background-color: whitesmoke;
-    }
-  }
-
-  .TokenAmount-token-modal {
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-
-    .TokenAmount-token-modal-header {
-      display: flex;
-      padding: 10px;
-      font-size: 18px;
-      border-bottom: 1px solid lightgray;
-      .TokenAmount-token-modal-header-image {
-        display: flex;
-        align-items: center;
-        margin-right: 5px;
-      }
-    }
-    .TokenAmount-token-modal-list {
-      flex: 1;
-      overflow-y: auto;
-      .TokenAmount-token {
-        display: flex;
-        border: 2px solid transparent;
-        border-radius: 12px;
-        font-size: 24px;
-        padding: 10px;
-        margin: 5px;
-
-        &:hover {
-          border: 2px solid #a30786;
-        }
-      }
-
-      .TokenAmount-token-active {
-        background-color: whitesmoke;
-      }
-    }
-
-    .TokenAmount-token-modal-footer {
-      border-top: 1px solid lightgray;
-      .TokenAmount-token-modal-button {
-        background-color: #00c3b4;
-        color: #ffffff;
-        padding: 10px 15px;
-        margin: 10px;
-        font-size: 18px;
-        border: 1px solid #00c3b4;
-        border-radius: 12px;
-        text-transform: uppercase;
-        font-weight: 600;
-
-        &:hover {
-          background-color: #00ab9e;
-        }
-      }
-    }
-  }
-
-  &:hover {
-    border: 2px solid #a30786 !important;
-  }
-}
-
-.TokenAmount-active {
-  border: 2px solid #a30786 !important;
-  font-weight: 700;
-}
-
-input[type='number']::-webkit-inner-spin-button,
-input[type='number']::-webkit-outer-spin-button {
-  -webkit-appearance: none;
-  margin: 0;
-}
