+++ frontend/src/components/TokenAmount/TokenAmount.scss
@@ -1,0 +1,59 @@
+.TokenAmount {
+  display: flex;
+  flex-direction: column;
+  border: 2px solid lightgray;
+  border-radius: 12px;
+  padding: 16px;
+
+  .TokenAmount-title {
+    margin: 0 0 12px 0;
+    font-size: 16px;
+    text-align: left;
+  }
+
+  .TokenAmount-content {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+  }
+
+  .TokenAmount-amount {
+    background-color: transparent;
+    font-size: 24px;
+    border: none;
+    outline: none;
+  }
+
+  .TokenAmount-token-select {
+    background-color: transparent;
+    background-repeat: no-repeat;
+    border: none;
+    border-radius: 12px;
+    outline: none;
+    cursor: pointer;
+    font-size: 16px;
+    font-weight: 600;
+    padding: 0 1rem;
+    height: 2.2rem;
+    user-select: none;
+
+    &:hover {
+      background-color: whitesmoke;
+    }
+  }
+
+  &:hover {
+    border: 2px solid #a30786 !important;
+  }
+}
+
+.TokenAmount-active {
+  border: 2px solid #a30786 !important;
+  font-weight: 700;
+}
+
+input[type='number']::-webkit-inner-spin-button,
+input[type='number']::-webkit-outer-spin-button {
+  -webkit-appearance: none;
+  margin: 0;
+}
