+++ frontend/src/components/TokenAmount/TokenModal.scss
@@ -1,0 +1,57 @@
+.TokenAmount-token-modal {
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+
+  .TokenAmount-token-modal-header {
+    display: flex;
+    padding: 10px;
+    font-size: 18px;
+    border-bottom: 1px solid lightgray;
+    .TokenAmount-token-modal-header-image {
+      display: flex;
+      align-items: center;
+      margin-right: 5px;
+    }
+  }
+  .TokenAmount-token-modal-list {
+    flex: 1;
+    overflow-y: auto;
+    .TokenAmount-token {
+      display: flex;
+      border: 2px solid transparent;
+      border-radius: 12px;
+      font-size: 24px;
+      padding: 10px;
+      margin: 5px;
+
+      &:hover {
+        border: 2px solid #a30786;
+      }
+    }
+
+    .TokenAmount-token-active {
+      background-color: whitesmoke;
+    }
+  }
+
+  .TokenAmount-token-modal-footer {
+    border-top: 1px solid lightgray;
+    .TokenAmount-token-modal-close-button {
+      background-color: #a30786;
+      color: #ffffff;
+      padding: 10px 15px;
+      margin: 10px;
+      font-size: 18px;
+      border: 1px solid #a30786;
+      border-radius: 12px;
+      text-transform: uppercase;
+      font-weight: 600;
+
+      &:hover {
+        background-color: #940678;
+      }
+    }
+  }
+}
