+++ frontend/src/components/common/WalletModal.scss
@@ -1,31 +1,31 @@
-.SwapComponent-wallet-modal {
-  height: 100%;
+.Wallet-modal {
   display: flex;
   flex-direction: column;
   justify-content: flex-start;
+  height: 100%;
 
-  .SwapComponent-wallet-modal-header {
+  .Wallet-modal-header {
     display: flex;
     padding: 10px;
     font-size: 18px;
     border-bottom: 1px solid lightgray;
     font-weight: 700;
-    .SwapComponent-wallet-modal-header-image {
+    .Wallet-modal-header-image {
       display: flex;
       align-items: center;
       margin-right: 5px;
     }
   }
 
-  .SwapComponent-wallet-modal-select {
+  .Wallet-modal-select {
     background-color: transparent;
     border: 2px solid lightgray;
     margin: 15px;
     border-radius: 12px;
     font-size: 18px;
-    font-weight: 500;
+    font-weight: 700;
 
-    .SwapComponent-wallet-modal-item {
+    .Wallet-modal-item {
       display: flex;
       justify-content: space-between;
       align-items: center;
